feat : add Not-found Middleware to prevent if the endpoint has been writing very well
