+++ .env
@@ -1,2 +1,2 @@
 POSTGRESQL_DB_URI = "postgres://wqakbgpi:v4OOv0HY6Z0AcebfXUMymuP2_6Lqimxr@rajje.db.elephantsql.com/wqakbgpi"
-PORT = 4000
+PORT = 4500
