+++ middleware/ErrorMiddleware.js
@@ -1,0 +1,7 @@
+const notFoundError = (req, res, next) => {
+  const error = new Error(`Not Found - ${req.originalUrl}`);
+  res.status(404);
+  next(error);
+};
+
+export default notFoundError;
